app/requirements.txt
black==23.3.0
Django==4.1.7
dj-database-url==1.3.0
djangorestframework==3.14.0
flake8==6.0.0
gunicorn==20.1.0
isort==5.12.0
psycopg2-binary==2.9.5
pytest-django==4.5.2
pytest==7.2.2
pytest-cov==4.0.0
# whitenoise==6.4.0
